Move convertCurrencyBySymbol logic into separate helper
