+++ workers/loc.api/sync/transaction.tax.report/index.js
@@ -9,7 +9,8 @@ const {
   getPubTradeChunkPayloads,
   TRX_TAX_STRATEGIES,
   reMapTrades,
-  reMapMovements
+  reMapMovements,
+  convertCurrencyBySymbol
 } = require('./helpers')
 
 const { decorateInjectable } = require('../../di/utils')
@@ -238,58 +239,7 @@ class TransactionTaxReport {
         pubTrades.push(...chunk)
       }
 
-      for (const trxDataItem of trxData) {
-        let lastIndex = 0
-
-        for (let i = lastIndex; pubTrades.length > i; i += 1) {
-          const pubTrade = pubTrades[i]
-          const isLastPubTrade = (i + 1) === pubTrades.length
-
-          if (
-            (
-              pubTrade?.mts > trxDataItem.trx.mtsCreate &&
-              !isLastPubTrade
-            ) ||
-            !Number.isFinite(pubTrade?.price) ||
-            pubTrade.price === 0
-          ) {
-            continue
-          }
-
-          lastIndex = i
-          trxDataItem.trx[trxDataItem.mainPricePropName] = pubTrade.price
-
-          if (trxDataItem.trx.isMovements) {
-            trxDataItem.trx.execPrice = pubTrade.price
-
-            break
-          }
-          if (
-            !Number.isFinite(trxDataItem.trx.execPrice) ||
-            trxDataItem.trx.execPrice === 0
-          ) {
-            break
-          }
-          if (
-            trxDataItem.isNotFirstSymbForex &&
-            !trxDataItem.isNotFirstSymbForex
-          ) {
-            trxDataItem.trx[trxDataItem.secondPricePropName] = (
-              pubTrade.price / trxDataItem.trx.execPrice
-            )
-          }
-          if (
-            !trxDataItem.isNotFirstSymbForex &&
-            trxDataItem.isNotFirstSymbForex
-          ) {
-            trxDataItem.trx[trxDataItem.secondPricePropName] = (
-              pubTrade.price * trxDataItem.trx.execPrice
-            )
-          }
-
-          break
-        }
-      }
+      convertCurrencyBySymbol(trxData, pubTrades)
     }
   }
